first version of FuzzyMedian on GPU

+++ AFTEREFFECT_PLUGIN.DEV/FuzzyMedianFilter/include/FuzzyRules.hpp
@@ -17,7 +17,7 @@
     const float& sqSigma
  ) noexcept
  {
-    return exp(-((d - m) * (d - m)) / (2.f * sqSigma));
+    return expf(-((d - m) * (d - m)) / (2.f * sqSigma));
  }
